Move SupportInfo component up to the Reading settings Card
